agent 05/requirements_data_analytics.txt
fastapi
uvicorn
python-multipart
python-dotenv
anthropic
openai
